==> packages/codcat/codcat-0.2.0.tar.gz/codcat-0.2.0/notebooks/2-preprocessing.ipynb <==
## Preprocessing

Source code preprocessing has several features:

1) Each language has its own grammar.
When working with source code, lexemes are usually individual words or code elements, such as variable names, function names, and operator names.
2) Source code preprocessing also requires certain steps that are unique to code, such as removing code comments, handling variable declarations and imports, and dealing with the peculiarities of coding languages.

For now we will choose `TweetTokenizer`, in the future we will describe the preprocessing with `code_tokenize` (tree-sitter) packets.

from nltk import NLTKWordTokenizer, TweetTokenizer

Removing some tokens (e.g., a string literal or a number) must be done at the tokenization stage, since only at this point do we know the types of tokens.

So in the following example, using `tokenize` removes `'bar'`

text = '''
def foo(bar: int) -> dict:
    return {'bar': bar}
'''

' '.join(NLTKWordTokenizer().tokenize(text))

' '.join(TweetTokenizer().tokenize(text))

text = '''
#include <iostream>
using namespace std;

int main() { double n1, n2, n3; cout << "Hello, world!"; }
'''

' '.join(NLTKWordTokenizer().tokenize(text))

' '.join(TweetTokenizer().tokenize(text))

## Save preprocessed data

import pandas as pd

train = pd.read_csv('train.csv')

test = pd.read_csv('test.csv')

tokenizer = TweetTokenizer()

train['code'] = train.apply(lambda x: ' '.join(tokenizer.tokenize(x['code'])), axis=1)

test['code'] = test.apply(lambda x: ' '.join(tokenizer.tokenize(x['code'])), axis=1)

train['code'] = train['code'].apply(lambda x: x.lower().strip())
test['code'] = test['code'].apply(lambda x: x.lower().strip())

train = train[(train['code'].map(len) > 0) & (train['code'] != 'null')]
test = test[(test['code'].map(len) > 0) & (train['code'] != 'null')]

train.to_csv('train-preprocessed.csv', index=None)
test.to_csv('test-preprocessed.csv', index=None)
