==> packages/codcat/codcat-0.2.0.tar.gz/codcat-0.2.0/notebooks/3-bagofwords-tfidf.ipynb <==
## Bag of words & TF-IDF

import pandas as pd
import numpy as np

train = pd.read_csv('train-preprocessed.csv')
test = pd.read_csv('test-preprocessed.csv')

## Tokenization

In this notebook I will do a little analysis of `Bag of Words`, `TF-IDF` with tokenization by words (by default) and with space-separated tokenization (because the pre-processed data contain exactly tokens with space-separated tokenization)

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report

y_train = train['language']
y_test = test['language']

vectorizer = CountVectorizer(max_df=0.7, min_df=0.003)
vectorizer_split = CountVectorizer(max_df=0.7, min_df=0.003, tokenizer=lambda x: x.split())

X_train_vectorized = vectorizer.fit_transform(train['code'])
X_test_vectorized = vectorizer.transform(test['code'])

mnb = MultinomialNB().fit(X_train_vectorized, y_train)
preds = mnb.predict(X_test_vectorized)
print(classification_report(y_test, preds, zero_division=0))

X_train_vectorized = vectorizer_split.fit_transform(train['code'])
X_test_vectorized = vectorizer_split.transform(test['code'])

mnb = MultinomialNB().fit(X_train_vectorized, y_train)
preds = mnb.predict(X_test_vectorized)
print(classification_report(y_test, preds, zero_division=0))

text = 'def foo ( bar ) : return bar'

vectorized = vectorizer_split.transform([text])

mnb.predict(vectorized)

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.ensemble import RandomForestClassifier

pipe = Pipeline(
    steps=[
        ('tfidf', TfidfVectorizer(token_pattern='\S+')),
        ('clf', RandomForestClassifier())
    ]
)

pipe.fit(train['code'], y_train)

preds = pipe.predict(test['code'])
print(classification_report(y_test, preds, zero_division=0))

## Stop words calculation try

https://www.researchgate.net/publication/318969652_AN_AUTO-GENERATED_APPROACH_OF_STOP_WORDS_USING_AGGREGATED_ANALYSIS

import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
import logging
logging.basicConfig(format='%(asctime)s : %(levelname)s : %(message)s', level=logging.INFO)

# Based on: https://github.com/bnriiitb/autostopwordgen

class AutoStopWordsGen:

    def __init__(self, corpus):
        """
        :param corpus: corpus
        """
        logging.info('initializing the AutoStopwordsGen started')
        cv = CountVectorizer(tokenizer=lambda x: x.split())
        cvft = cv.fit_transform(corpus)
        tfidfcv = TfidfVectorizer(tokenizer=lambda x: x.split())
        tfidfcvft = tfidfcv.fit_transform(corpus)
        self.tfidfcv=tfidfcv
        self.cvft=cvft
        logging.info('initializing the AutoStopwordsGen completed')

    def get_stopwords(self, top_n=.95, last_n=.1):

        """
        :param top_n: top n percent threshold
        :param last_n: last n percent threshold
        :return: returns stopwords
        """
        logging.info('generating stopwords started')
        word_freq_df = pd.DataFrame({'word': self.tfidfcv.get_feature_names_out(),
                                     'frequency':np.asarray(self.cvft.sum(axis=0)).ravel().tolist(),
                                     'idf':self.tfidfcv.idf_})
        word_freq_df.sort_values(by=['frequency'],ascending = False,inplace=True)

        word_freq_df['prob'] = word_freq_df.frequency/word_freq_df.shape[0]
        word_freq_df['entropy'] = word_freq_df.prob.apply(lambda x: x*np.log(1/x))
        word_freq_df['vp'] = np.power(word_freq_df.prob-word_freq_df.prob.mean(),2)/word_freq_df.shape[0]

        stopwords=dict({'frequency':[],'idf':[],'entropy':[],'vp':[]})
        cols=['frequency','entropy','vp', 'idf']

        for col in cols:
            # print(col,' : ',word_freq_df[col].quantile([last_n,top_n]).tolist())
            if(col=='frequency'):
                top_5_percent=word_freq_df[col].quantile([last_n,top_n]).tolist()[1]
                stopwords[col]=word_freq_df[word_freq_df[col]>=top_5_percent].word.tolist()
            else:
                last_10_percent=word_freq_df[col].quantile([last_n,top_n]).tolist()[0]
                stopwords[col]=word_freq_df[word_freq_df[col]<=last_10_percent].word.tolist()

        for key in stopwords.keys():
            stopwords[key]=set(stopwords[key])
            
        very_high_aggregation = word_freq_df[word_freq_df.frequency<2].word.tolist()
        
        very_high_aggregation.extend(list(stopwords['frequency'].intersection(stopwords['entropy']).intersection(stopwords['vp'])))
        very_high_aggregation=list(set(very_high_aggregation))
        logging.info('# stopwords generated :: '+str(len(very_high_aggregation)))
        logging.info('generating stopwords completed')
        return very_high_aggregation

asw = AutoStopWordsGen(train['code'])

stop_words = asw.get_stopwords()

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report

vectorizer = TfidfVectorizer(tokenizer=lambda x: x.split(), stop_words=stop_words)
X_train_vectorized = vectorizer.fit_transform(train['code'])
X_test_vectorized = vectorizer.transform(test['code'])
clf = RandomForestClassifier(n_jobs=-1, random_state=42).fit(X_train_vectorized, y_train)
preds = clf.predict(X_test_vectorized)
print(classification_report(y_test, preds, zero_division=0))

From this we can conclude that deleting `54404` words with this algorithm can reduce the number of words in the corpus and the quality will remain virtually unchanged. Therefore, in the future you can try to apply this list of stop words.

import json

with open('stop-words.json', 'w') as f:
    f.write(json.dumps([w.strip() for w in stop_words]))
