==> packages/codcat/codcat-0.2.0.tar.gz/codcat-0.2.0/notebooks/5-sentence-transformers.ipynb <==
# Sentence-transformers

Sentence-transformers are a type of natural language processing (NLP) model that is trained to convert a given sentence or text into a vector representation that captures the semantic meaning of the text. These vector representations can then be used for a variety of tasks such as text classification, semantic similarity, and text clustering, among others.

`microsoft/codebert-base` is a pre-trained language model that was specifically trained on source code tasks. It is a variant of the original BERT model, which was pre-trained on a large corpus of text data. The CodeBERT-base model was pre-trained on a massive dataset of 6 million lines of code, which was sourced from various open-source software projects.

The reason for using Microsoft/CodeBERT-base as the base model for sentence-transformers is that it has several advantages over other pre-trained language models. 

Firstly, as it was trained on source code tasks, it has a better understanding of the syntactical and structural components of code, which makes it more suitable for tasks related to code analysis and code generation. This is particularly useful for developers who are looking to automate certain parts of their coding process.

Secondly, `microsoft/codebert-base` has been fine-tuned on a variety of programming-related tasks, such as code completion, code summarization, and code search. This means that it has a better understanding of the programming domain and can provide more accurate and relevant results for programming-related tasks.

Finally, `microsoft/codebert-base` has been shown to perform well on a variety of downstream NLP tasks, such as sentence classification and semantic similarity. This is due to its ability to capture the semantic meaning of the text, which is essential for many NLP tasks.

from sentence_transformers import SentenceTransformer

model = SentenceTransformer('microsoft/codebert-base')

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier

train = pd.read_csv('train-preprocessed.csv')
test = pd.read_csv('test-preprocessed.csv')

X_train_embeddings = model.encode(train['code'].values)

X_test_embeddings = model.encode(test['code'].values)

from lightgbm import LGBMClassifier

clf = LGBMClassifier(random_state=42).fit(X_train_embeddings, train['language'])

from sklearn.metrics import classification_report

print(classification_report(test['language'], clf.predict(X_test_embeddings)))
