==> packages/codcat/codcat-0.2.0.tar.gz/codcat-0.2.0/notebooks/6-code25-dataset.ipynb <==
## Full SCC dataset

In the previous steps, we used a dataset that was better pre-cleaned and contained enough data. However, for the purity of the experiment, we should use the original dataset from the SCC article, divide it in the necessary ratio, and make measurements on it. 

Link for dataset: https://github.com/Kamel773/SourceCodeClassification

We downloaded into `code25` folder.

from codcat.utils.dataset import load_labeled_data
from sklearn.model_selection import train_test_split

data = load_labeled_data('/home/qnbhd/Development/sniplang/code25')

In SCC, SCC++, DeepSCC papers authors use 20% for testing

train, test = train_test_split(data, random_state=42, test_size=0.2)

train.to_json('train-code25.json')
test.to_json('test-code25.json')

To test the basic models, we will further reduce the size of the training and test samples.

_, temp = train_test_split(test, random_state=42, test_size=0.3)
train_smol, test_smol = train_test_split(temp, random_state=42, test_size=0.2)

train_smol.to_json('train-code25-smol.json')
test_smol.to_json('test-code25-smol.json')

train_smol.shape

test_smol.shape

train_smol['language'].value_counts().sort_index()

test_smol['language'].value_counts().sort_index()
