==> packages/codcat/codcat-0.2.0.tar.gz/codcat-0.2.0/notebooks/7-code-tokenize.ipynb <==
## Preprocessing with code-tokenize & Pygments

The code_tokenize package was selected to tokenize the source code during the training phase. It is based on the tree-sitter package, which has tokenizers for a large number of programming languages and more.

- code_tokenize: https://github.com/cedricrupb/code_tokenize
- tree-sitter: https://tree-sitter.github.io/tree-sitter/

We would like to point out that it will only be used in the TRAINING part because of the need to know the marker.

The idea is simple: to clear dataset not from words, but from stop-types. This is a more suitable approach, since programming languages have their own structure. For example, we can remove comments or some literals.

!pip install code_tokenize pygments

from codcat.tok import preprocess

### Let's check preprocessing for some languages

print(preprocess('''
# This is a Bash comment.
echo "This is Code" # This is an inline Bash comment.
a=5
b=7
''', lang='bash'))

Vb.net is an exception, because it is not supported by the tree-sitter package. The text is returned in its original form.

print(preprocess('''
' This is a comment beginning at the left edge of the screen.
text1.Text = "Hi!"   ' This is an inline comment.
''', lang='vb.net'))

print(preprocess('''
# this is a print statement
# it prints Hello World

print("Hello World") 
''', lang='r'))

print(preprocess('''
// some c++ comment
/* Multi
Line 
Comment */

int main() {
    return 0;
}
''', lang='cpp'))

print(preprocess('''
// some c comment
/* Multi
Line 
Comment */

int main() {
    return 0;
}
''', lang='c'))

print(preprocess('''
#!/usr/bin/ruby -w
# This is a single line comment.

puts "Hello, Ruby!"
''', lang='ruby'))

print(preprocess('''
// This is a comment
System.out.println("Hello World");
''', lang='java'))

print(preprocess('''
<?php
    echo 'This is a test'; // This is a one-line c++ style comment
    /* This is a multi line comment
       yet another line of comment */
    echo 'This is yet another test';
    echo 'One Final Test'; # This is a one-line shell-style comment
?>
''', lang='php'))

By default, only comments are deleted. To evaluate, you can try deleting different types in the source code.

In practice, this method does not always improve the quality of classification, but it can improve the stability of classification. Comments in present languages exist everywhere, but they do not play a large role in determining the final language.

## Pygments

Using pygments is a good approach in getting rid of little-playing tokens in the text. It can also be used to remove comments or certain literals. 
The approach suggested below allows you not to use knowledge of the test data to clean up.

In general, the structure of the approach is as follows:

`Data -> Initial Clf -> Pygments -> Preprocessed Data -> Clf`

That is, we first train a simple enough classifier or use a pre-trained one to predict the class label with good accuracy. Then we use these predictions to run the lexical analyzer, do preprocessing. And then we use the approaches to classify the text. 

This approach does not always show good results, for example on the Code25 dataset (SCC) it degrades quality, and on the original dataset it neither degrades nor improves.

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import Pipeline
import lime
import lime.lime_text
from sklearn.metrics import classification_report
from codcat.pyg import get_lexer, Prelude
from nltk import TweetTokenizer

train = pd.read_json('train-code25-smol.json')
test = pd.read_json('test-code25-smol.json')

model = Pipeline(
    [
        ('cv', TfidfVectorizer()),
        ('rf', MultinomialNB()),
    ]
)

model.fit(train['code'], train['language'])

print(classification_report(test['language'], model.predict(test['code'])))

model_prelude = Pipeline(
    [
        ('prelude', Prelude(model)),
        ('cv', TfidfVectorizer()),
        ('rf', MultinomialNB()),
    ]
)

model_prelude.fit(train['code'], train['language'])

print(classification_report(test['language'], model_prelude.predict(test['code'])))
